+++ main/src/ca/uwaterloo/flix/language/phase/unification/FormulaUnification.scala
@@ -18,7 +18,7 @@ package ca.uwaterloo.flix.language.phase.unification
 import ca.uwaterloo.flix.api.Flix
 import ca.uwaterloo.flix.language.ast._
 import ca.uwaterloo.flix.language.phase.unification.ExplicitFormula.{VarOrEff, fromBoolType, toType}
-import ca.uwaterloo.flix.language.phase.unification.BoolAlgebraTable.minimizeFormula
+import ca.uwaterloo.flix.language.phase.unification.BoolFormulaTable.minimizeFormula
 import ca.uwaterloo.flix.util.Result.Ok
 import ca.uwaterloo.flix.util.collection.Bimap
 import ca.uwaterloo.flix.util.{InternalCompilerException, Result}
